requirements.txt
# Core Frameworks
langgraph>=0.0.30
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Storage & Databases
redis>=5.0.0
chromadb>=0.4.20
overrides==7.3.1

# Utilities
tiktoken>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tabulate>=0.9.0
